refactor: Extend Component trait with actions
fn input now maps to an Action, which is handled by
fn handle_action. This makes it possible to configure
which key events maps to which actions, and to have
a global menu generate actions in addition to shortcut keys.

By adding handle_action to the Component trait, the compiler
will check that this is the case for all Components.

In order to implement this refactoring gradually, a 
default implementation of handle_action is provided
which does nothing.